:art: #5 Page,Counter,Clockを別々のContainerに分割
