+++ components/page.tsx
@@ -1,57 +1,40 @@
 import Link from 'next/link';
 
-import Counter from './Counter';
-import Clock from './Clock';
+import Counter from '../container/Counter';
+import Clock from '../container/Clock';
+import { FC } from 'react';
 
 export interface PageProps {
-  error?: any;
-  linkTo?: any;
-  NavigateTo?: any;
-  placeholderData?: any;
   title?: string;
-  lastUpdate?: any;
-  light?: any;
-  count?: any;
-  increment?: any;
-  decrement?: any;
-  reset?: any;
+  linkTo?: string;
+  NavigateTo?: string;
+  placeholderData?: any;
+  error?: any;
 }
 
-function Page({
-  error,
-  linkTo,
-  NavigateTo,
-  placeholderData,
-  title,
-  lastUpdate,
-  light,
-  count,
-  increment,
-  decrement,
-  reset,
-}: PageProps) {
-  return (
-    <div>
-      <h1>{title}</h1>
-      <Clock lastUpdate={lastUpdate} light={light} />
-      <Counter
-        count={count}
-        increment={increment}
-        decrement={decrement}
-        reset={reset}
-      />
-      <nav>
-        <Link href={linkTo}>
-          <a>Navigate: {NavigateTo}</a>
-        </Link>
-      </nav>
-      {placeholderData && (
-        <pre>
-          <code>{JSON.stringify(placeholderData, null, 2)}</code>
-        </pre>
-      )}
-      {error && <p style={{ color: 'red' }}>Error: {error.message}</p>}
-    </div>
-  );
-}
+const Page: FC<PageProps> = ({
+  linkTo = '',
+  NavigateTo = '',
+  placeholderData = {},
+  title = '',
+  error = {},
+}) => (
+  <div>
+    <h1>{title}</h1>
+    <Clock />
+    <Counter />
+    <nav>
+      <Link href={linkTo}>
+        <a>Navigate: {NavigateTo}</a>
+      </Link>
+    </nav>
+    {placeholderData && (
+      <pre>
+        <code>{JSON.stringify(placeholderData, null, 2)}</code>
+      </pre>
+    )}
+    {error && <p style={{ color: 'red' }}>Error: {error.message}</p>}
+  </div>
+);
+
 export default Page;
