+++ reducers/rootReducer.ts
@@ -1,8 +1,14 @@
 import { combineReducers } from 'redux';
 
-import clockReducer from './clock';
-import counterReducer from './counter';
-import userDataReducer from './userData';
+import clockReducer, { ClockState } from './clock';
+import counterReducer, { CounterState } from './counter';
+import userDataReducer, { UserDataState } from './userData';
+
+export interface ReduxState {
+  clock: ClockState;
+  counter: CounterState;
+  userData: UserDataState;
+}
 
 const rootReducer = combineReducers({
   clock: clockReducer,
